+++ sso_front/src/__mocks__/handlers.ts
@@ -1,5 +1,0 @@
-// import { rest } from "msw";
-
-export const handlers = [];
-
-export const errorHandlers = [];