+++ web/usuaris/src/components/molecules/index.ts
@@ -1,0 +1,2 @@
+export { DateRange, type TDateRange } from './DateRange'
+export { Table, type TTable } from './Table'