+++ web/usuaris/src/constants/urls.ts
@@ -1,0 +1,3 @@
+export const urls = {
+  getUsers: '/api/v1/dashboard/users',
+}