+++ web/usuaris/src/helpers/index.ts
@@ -1,0 +1,1 @@
+export { getUsers } from './fetchers'